fix float - int

+++ pkg/models/models.go
@@ -19,8 +19,8 @@ type WithdrawInfo struct {
 }
 
 type Withdraw struct {
-	Order string `json:"order"`
-	Sum   int    `json:"sum"`
+	Order string  `json:"order"`
+	Sum   float32 `json:"sum"`
 }
 
 type AuthModel struct {
